Added disableDeleteKey config prop (#273)

+++ src/ImageAnnotator.jsx
@@ -80,12 +80,16 @@ export default class ImageAnnotator extends Component  {
         this.props.onCancelSelected(selectedAnnotation);
       }
     } else if (evt.which === 46) { // Delete
-      const { selectedAnnotation } = this.state;
-      if (selectedAnnotation) {
-        if (selectedAnnotation.isSelection) {
-          this.onCancelAnnotation(selectedAnnotation);
-        } else {
-          this.onDeleteAnnotation(selectedAnnotation);
+      const { disableDeleteKey } = this.props.config;
+
+      if (!disableDeleteKey) {
+        const { selectedAnnotation } = this.state;
+        if (selectedAnnotation) {
+          if (selectedAnnotation.isSelection) {
+            this.onCancelAnnotation(selectedAnnotation);
+          } else {
+            this.onDeleteAnnotation(selectedAnnotation);
+          }
         }
       }
     }
